refactor: Separate settings handling into dedicated class
- Create new `S3_Media_Sync_Settings` class to manage plugin settings
- Move settings rendering and validation methods from `S3_Media_Sync` to new class
- Update main plugin file to use new settings handler
- Modify test cases to work with new settings management approach
- Simplify plugin initialization and settings loading
