+++ s3-media-sync.php
@@ -9,17 +9,17 @@
  * Version: 1.4.1
  */
 
- const S3_MEDIA_SYNC_FILE = __FILE__;
+// Define plugin constants
+define( 'S3_MEDIA_SYNC_FILE', __FILE__ );
+
+// Load required files
+require_once dirname( __FILE__ ) . '/inc/class-s3-media-sync-settings.php';
+require_once dirname( __FILE__ ) . '/inc/class-s3-media-sync.php';
 
 /**
  * Class that handles s3 stream wrapper
  */
 require_once plugin_dir_path( __FILE__ ) . 'inc/class-s3-media-sync-stream-wrapper.php';
-
-/**
- * Class that handles core plugin functionality
- */
-require_once plugin_dir_path( __FILE__ ) . 'inc/class-s3-media-sync.php';
 
 /**
  * Class that adds WP-CLI commands
@@ -28,18 +28,5 @@ if ( defined( 'WP_CLI' ) && WP_CLI && class_exists( 'WPCOM_VIP_CLI_Command' ) ) 
 	require_once dirname( __FILE__ ) . '/inc/class-s3-media-sync-wp-cli.php';
 }
 
-/**
- * Set up the plugin
- */
-function s3_media_sync_setup() {
-
-	// Ensure the AWS SDK can be loaded.
-	if ( ! class_exists( '\\Aws\\S3\\S3Client' ) ) {
-		// Require AWS Autoloader file.
-		require_once dirname( __FILE__ ) . '/vendor/autoload.php';
-	}
-
-	$instance = S3_Media_Sync::init();
-	$instance->setup();
-}
-add_action( 'plugins_loaded', 's3_media_sync_setup' );
+// Initialize the plugin
+add_action( 'plugins_loaded', [ S3_Media_Sync::init(), 'setup' ] );
