+++ tests/Integration/ErrorHandlingTest.php
@@ -17,8 +17,9 @@ use S3_Media_Sync\Tests\TestCase;
  *
  * @group integration
  * @group error-handling
- * @covers S3_Media_Sync
- * @uses S3_Media_Sync_Stream_Wrapper
+ * @covers \S3_Media_Sync
+ * @uses \S3_Media_Sync_Stream_Wrapper
+ * @uses \S3_Media_Sync_Settings
  */
 class ErrorHandlingTest extends TestCase {
