+++ tests/Integration/HooksTest.php
@@ -15,8 +15,9 @@ use S3_Media_Sync\Tests\TestCase;
  *
  * @group integration
  * @group hooks
- * @covers S3_Media_Sync
- * @uses S3_Media_Sync_Stream_Wrapper
+ * @covers \S3_Media_Sync
+ * @uses \S3_Media_Sync_Stream_Wrapper
+ * @uses \S3_Media_Sync_Settings
  */
 class HooksTest extends TestCase {
