+++ tests/Integration/SettingsTest.php
@@ -17,8 +17,9 @@ use S3_Media_Sync\Tests\TestCase;
  *
  * @group integration
  * @group settings
- * @covers S3_Media_Sync
- * @uses S3_Media_Sync_Stream_Wrapper
+ * @covers \S3_Media_Sync_Settings
+ * @uses \S3_Media_Sync
+ * @uses \S3_Media_Sync_Stream_Wrapper
  */
 class SettingsTest extends TestCase {
 
@@ -87,15 +88,9 @@ class SettingsTest extends TestCase {
 			->with( $settings['bucket'] )
 			->andReturn( false );
 
-		// Update WordPress option
+		// Update WordPress option and settings handler
 		update_option( 's3_media_sync_settings', $settings );
-
-		$this::set_private_property(
-			$this->s3_media_sync::class,
-			$this->s3_media_sync,
-			'settings',
-			$settings
-		);
+		$this->settings_handler->update_settings( $settings );
 
 		// Only set the S3 client if we should validate
 		if ( $should_validate ) {
@@ -112,7 +107,7 @@ class SettingsTest extends TestCase {
 		// Clear any errors that might have been set during setup
 		$wp_settings_errors = [];
 
-		$this->s3_media_sync->s3_media_sync_settings_validation( $settings );
+		$this->settings_handler->settings_validation( $settings );
 
 		$admin_error_codes = wp_list_pluck( get_settings_errors(), 'code' );
 
@@ -137,20 +132,12 @@ class SettingsTest extends TestCase {
 	 * @throws \ReflectionException If reflection fails.
 	 */
 	public function test_settings_are_saved( array $settings, string $error_code, bool $should_validate ): void {
-		$this::set_private_property(
-			$this->s3_media_sync::class,
-			$this->s3_media_sync,
-			'settings',
-			$settings
-		);
+		// Update settings
+		$this->settings_handler->update_settings( $settings );
 
 		$this->s3_media_sync->setup();
 
-		$saved_settings = $this::get_private_property(
-			$this->s3_media_sync::class,
-			$this->s3_media_sync,
-			'settings'
-		);
+		$saved_settings = $this->settings_handler->get_settings();
 
 		Assert::assertSame( $settings, $saved_settings );
 	}
