+++ tests/Integration/StreamWrapperTest.php
@@ -21,8 +21,9 @@ use S3_Media_Sync\Tests\TestCase;
  *
  * @group integration
  * @group stream-wrapper
- * @covers S3_Media_Sync
- * @covers S3_Media_Sync_Stream_Wrapper
+ * @covers \S3_Media_Sync_Stream_Wrapper
+ * @uses \S3_Media_Sync
+ * @uses \S3_Media_Sync_Settings
  */
 class StreamWrapperTest extends TestCase {
